More Gallery Navigation
Set a fixed height for gallery main so the navigation would stop moving around. Changed the padding on nav icons so they wouldn't squash when the border is added.  Added the gallery of works on paper. Created scaled down versions of the images for the icons to greatly reduce load times and lagginess.