Archive and Extras
Added the CSS to the archive tabs and grid. Added all the archive images and made them look nice. Renamed the Catalogue tag to News. Added the experiemental background to the home page. Tried making a modal for the archive images, but it was only for a singular image, not a gallery. I'll make it work for a gallery next time.