Gina Meeting
Had a meeting with Gina on Spet 30 and made lots of changes based on feedback. Moved around some images, and added/deleted others. Made the title and nav text capitilized and more spaced out, and changed the colors. Removed the background image. Changed the names of some tabs. Have to still add icons for new images and fix the ordering.