Email Updates
Removed a bunch of images from the archive. Moved some from archive to paintings and works on paper, and created their icons. Changed the tabs in the archive to just scrolling down. Edited some spacing in archive. Added Gina's CV in About and changed the h3 color. Commented out a bunch of unnecessary code. Removed the Publications tab.