Fuck php
Got the php working for the gallery and nav in paintings. The gallery is now read from the images folder, and the icons from the icon folder. They appear in the same order as the directory, which means however the fuck Windows 10 decides to sort them. Right now you only get to the php file from the index file. Clicking paintings from any other page brings you to the html file. Started work on the image information. Got the styling done. Still need to write the script, which should just be a few lines in gallery script.js. Renamed the "current work" folder to "paintings" so it actually makes sense. Added an info folder and created text files for each image. The php file tries to read the text files right now and fails, so this is a buggy push, but I gotta go to sleep. Also set up a local server with Apache and installed php on it on my computer, in order to test my php scripts. I have learned that fuck php.