Archive Modal + More Press
Wrote the archive modal script to open the modal when an image is clicked, and click through the images and information captions. It should work the same as the other scripts with the php, once that's done. Styled the modal in the html code, since even though I'll replace it with php, it's a lot easier to rapidly iterate and check the styling. Replaced Duet with a lighter, huge version. Added the catalogue links in a styled grid to the press page, with their own thumbnails. Started adding and styling the review links. Still need to comment the new code.