Comments + Press
Commented the new archive modal script, and edited the comments in the other two scripts. Added comments in the style sheet and deleted some commented-out code. Added the rest of the article links and the last catalogue, with a new thumbnail. Need to upload this to Gina to see what she thinks, and ask about the copy pasted articles from her old site.