Archive Info + Untested Php Conversion
Added all the information text files for the archive images. Created new php files for works on paper and the archive. Just copy pasted the same code and changed the directory path, and added a count for archive. Haven't tested them, and I think the way archive is set up, with the two folders, might create an error with the php, since I have to call the code twice. I might have to rename variables for it to work properly.