Press Collapsible
Added a collapsible article to the reviews section on the press page for articles that don't have links. Formatted the collapsibles and added text and images to them. Wrote a script to hide and show the collapsible. Made it so clicking on an html page while on that page goes to it's php page and vice versa for testing purposes. Tested all the php from the last two commits.