Special Projects + Wide
Renamed Public Art to Special Projects. Added images and icons for the rug project. Waiting on the guy to put it in the gallery. Added empty info for the special projects gallery so the script doesn't throw errors. I could've made an if statement but this was quicker. Also made a seperate directory for the wide images, icons, and infos in works on paper and made and styled a class to make them smaller. They didn't need to all be in a separate folder, but it was too risky due to the way Windows sorts them. Edited the php to accomidate the images, icons, and infos in the wide directory.