Bluehost Fix + Optimization
Big changes. Deleted my .htaccess file, and copied the Bluehost htaccess file as a backup.  Changed the image and icon php code in paintings, paper, and archive to work properly on Bluehost. Instead of a bunch a variable and reading dirs and such, I just glob all the files and spit it out. It works a lot better. Added a temporary txt file with the old code in case I need to refer to it. Reduced the size of all the images on the site to under 2 MB. Finished the new introduction in the README guide. Now I just need to wait for any feedback Gina has and that one rug photo and the site is done.