Gallery Navigation Update
Changed the way the navigation icons work for the image galleries. The nav bar will show a set of six images, and clicking the navigation arrows with switch between the sets of six images. Now it will switch sets while scrolling through images using the main gallery overlay. Created a new navigation index, code to determine it, and a helper function to display the correct set of icons. Created a helper function to switch navigation sets. Added the new functions to the gallery html. Edited the styling to stop the icons from being cut off.