More exhibition changes
Removed the exhibition notice from the front page. Removed the studio images from the Prince exhibition and added the remaining image info. Removed the heading from the Oakland exhibition and corrected some typos. Moved a studio image to the about page and added a max-width to it. Updated the client's CV on the about page.