Slideshow update
Updated the slideshow with new images and removed the old images. Added titles to the slideshow images with image information and compressed the images. Moved the deleted images from the Paintings and Works on Paper pages to the Archive page, along with their info, and deleted their icons. There is a known issue where wide image get stretched. I just need to copy the wide code from paintings into archive and move the wide images and info to a wide folder in archive, but I'll only do that if the client notices, as I have other work I need to do.