Gallery re-rework
Corrected some errors involving wide images in the slideshow and galleries. Added new images to the slideshow and deleted old ones. Deleted old paintings and added new ones with info and icons. Moved the works on paper images to a new archive section and added new works on paper images to the gallery with info and icons, and wide images. Added installation photos to the archive. Corrected some image info text. Added a "new" bold text notification to bring users to the new review in the navigation bar, the Prince Street exhibition, and next to the review on the press page. Updated the sitemap and added a google dns txt entry for the google search console.